move snapshot-version from core to model
