+++ Firestore/core/src/firebase/firestore/model/CMakeLists.txt
@@ -22,6 +22,8 @@ cc_library(
     field_path.h
     field_value.cc
     field_value.h
+    snapshot_version.cc
+    snapshot_version.h
     timestamp.cc
     timestamp.h
   DEPENDS
