+++ Firestore/core/test/firebase/firestore/core/CMakeLists.txt
@@ -16,7 +16,6 @@ cc_test(
   firebase_firestore_core_test
   SOURCES
     database_info_test.cc
-    snapshot_version_test.cc
     target_id_generator_test.cc
   DEPENDS
     firebase_firestore_core
