+++ Firestore/core/test/firebase/firestore/model/CMakeLists.txt
@@ -18,6 +18,7 @@ cc_test(
     database_id_test.cc
     field_path_test.cc
     field_value_test.cc
+    snapshot_version_test.cc
     timestamp_test.cc
     resource_path_test.cc
   DEPENDS
